requirements.txt
pymongo
Flask
fastapi
typing
uvicorn
python-dotenv